src/main/java/nl/ordina/jobcrawler/service/RoleService.java
package nl.ordina.jobcrawler.service;

import nl.ordina.jobcrawler.exception.RoleNotFoundException;
import nl.ordina.jobcrawler.model.Role;
import nl.ordina.jobcrawler.util.RoleName;
import nl.ordina.jobcrawler.repo.RoleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class RoleService {

    private final RoleRepository roleRepository;

    @Autowired
    public RoleService(RoleRepository roleRepository) {
        this.roleRepository = roleRepository;
    }

    public List<Role> findAll() {
        return roleRepository.findAll();
    }

    public Role save(Role role) {
        return roleRepository.save(role);
    }

    public boolean delete(Long id) {
        try {
            roleRepository.deleteById(id);
            return true;
        } catch (EmptyResultDataAccessException e) {
            throw new RoleNotFoundException(String.format("Role with id: %d not found.", id));
        }
    }

    public Optional<Role> findByName(RoleName roleName) {
        return roleRepository.findByName(roleName);
    }
}
